move globalStyles folder
